extractor/requirements.txt
# Deal-pipeline extractor dependencies (Task 1 — PDF read + Claude extraction + normalization)
# A minimal subset of the Nordic repo's requirements.txt. FastAPI/uvicorn land
# in Task 5; beautifulsoup4 + requests + openpyxl aren't used in the UK path.

anthropic>=0.18.0
pypdf>=4.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
